Options parsing with regex, universal char getter, char owner permission
These 3 features are linked - we now parse options and arguments with a regex which
detects multiple word arguments in quotation marks. This in turn allows us to use a simple
getter to get a character from a mention, which we use in our new char owner permission.
Overall this commit simplifies individual commands' code and is gonna allow us to expand
how we parse a character from an argument.